fixed imports, promoted optimise to general nonlinear
