+++ autofit/non_linear/nest/abstract_nest.py
@@ -1,19 +1,16 @@
-from typing import Optional, Tuple
+from typing import Optional
 
 import numpy as np
 from sqlalchemy.orm import Session
 
 from autoconf import conf
 from autofit import exc
-from autofit.graphical import ModelFactor, EPMeanField, MeanField, NormalMessage
-from autofit.graphical.utils import Status
 from autofit.non_linear.abstract_search import IntervalCounter
 from autofit.non_linear.abstract_search import NonLinearSearch
 from autofit.non_linear.initializer import InitializerPrior
-from autofit.non_linear.optimize.abstract_optimize import AbstractOptimizer
 
 
-class AbstractNest(NonLinearSearch, AbstractOptimizer):
+class AbstractNest(NonLinearSearch):
     def __init__(
             self,
             name=None,
@@ -53,81 +50,6 @@ class AbstractNest(NonLinearSearch, AbstractOptimizer):
             session=session,
             **kwargs
         )
-
-    def optimise(
-            self,
-            factor: ModelFactor,
-            model_approx: EPMeanField,
-            status: Optional[Status] = None
-    ) -> Tuple[EPMeanField, Status]:
-        """
-        Perform optimisation for expectation propagation. Currently only
-        applicable for ModelFactors created by the declarative interface.
-
-        1. Analysis and model classes are extracted from the factor.
-        2. Priors are updated from the mean field.
-        3. Analysis and model are fit as usual.
-        4. A new mean field is constructed with the (posterior) 'linking' priors.
-        5. Projection is performed to produce an updated EPMeanField object.
-
-        Parameters
-        ----------
-        factor
-            A factor comprising a model and an analysis
-        model_approx
-            A collection of messages defining the current best approximation to
-            some global model
-        status
-
-        Returns
-        -------
-        An updated approximation to the model having performed optimisation on
-        a single factor.
-        """
-
-        _ = status
-        if not isinstance(
-                factor,
-                ModelFactor
-        ):
-            raise NotImplementedError(
-                f"Optimizer {self.__class__.__name__} can only be applied to ModelFactors"
-            )
-
-        factor_approx = model_approx.factor_approximation(
-            factor
-        )
-        arguments = {
-            prior: factor_approx.model_dist[
-                prior
-            ].as_prior()
-            for prior in factor_approx.variables
-        }
-
-        model = factor.prior_model.mapper_from_prior_arguments(
-            arguments
-        )
-        analysis = factor.analysis
-
-        result = self.fit(
-            model=model,
-            analysis=analysis
-        )
-
-        new_model_dist = MeanField({
-            prior: NormalMessage.from_prior(
-                result.model.prior_with_id(
-                    prior.id
-                )
-            )
-            for prior in factor_approx.variables
-        })
-
-        projection, status = factor_approx.project(
-            new_model_dist,
-            delta=1
-        )
-        return model_approx.project(projection, status)
 
     class Fitness(NonLinearSearch.Fitness):
         def __init__(
